Improved HookRequest flexibility (const'd arguments)
git-svn-id: file:///home/nexis/Documents/luna-mirror/trunk@146 97ece4ad-cae2-4924-842c-0ef94757c1ff

+++ indra/newview/llagent.cpp
@@ -458,7 +458,7 @@ void LLAgent::init()
 	// Run a lua hook
 	std::string lolname;
 	this->getName(lolname);
-	LUA_CALL("OnAgentInit") << lolname << (int)isGodlike() << LUA_END;
+	LUA_CALL("OnAgentInit") << lolname << isGodlike() << LUA_END;
 }
 
 //-----------------------------------------------------------------------------
